refactor(providers): split ModelRegistry into per-concern peer files
internal/providers/registry.go was 1902 lines mixing seven concerns
(registration, init/refresh, cache load+save, metadata enrichment,
lookup, categories, runtime snapshots). Split into peer files in
the same package:

  registry.go (808)            struct, registration, lookup,
                               ListModels, categories, runtime snapshots,
                               provider-name/type indirection, splitModelSelector
  registry_init.go (494)       Initialize/Refresh/InitializeAsync,
                               fetchProviderInventory, refresh semaphore,
                               StartBackgroundRefresh, RefreshModelList,
                               provider runtime updates
  registry_cache.go (227)      LoadFromCache, SaveToCache
  registry_metadata.go (417)   SetModelList, EnrichModels, ResolveMetadata,
                               GetModelMetadata, ResolvePricing, snapshot
                               helpers, applyConfigMetadataOverrides,
                               enrichProviderModelMaps, registryAccessor

Pure relocation; tests pass.
