+++ internal/providers/registry.go
@@ -2,14 +2,8 @@
 package providers
 
 import (
-	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"log/slog"
-	"maps"
-	"net/http"
-	"reflect"
 	"slices"
 	"sort"
 	"strings"
@@ -197,558 +191,8 @@ func (r *ModelRegistry) RegisterProviderWithNameAndType(provider core.Provider, 
 	r.providerRuntime[providerName] = state
 }
 
-// Initialize fetches models from all registered providers and populates the registry.
-// This should be called on application startup.
-func (r *ModelRegistry) Initialize(ctx context.Context) error {
-	release, err := r.acquireRefresh(ctx)
-	if err != nil {
-		return err
-	}
-	defer release()
-	return r.initialize(ctx)
-}
 
-func (r *ModelRegistry) initialize(ctx context.Context) error {
-	// Get a snapshot of providers with a read lock
-	r.mu.RLock()
-	providers := make([]core.Provider, len(r.providers))
-	copy(providers, r.providers)
-	r.mu.RUnlock()
 
-	// Build new model maps without holding the lock.
-	// This allows concurrent reads to continue using the existing map
-	// while we fetch models from providers (which may involve network calls).
-	newModels := make(map[string]*ModelInfo)
-	newModelsByProvider := make(map[string]map[string]*ModelInfo)
-	var totalModels int
-	var failedProviders int
-	runtimeUpdates := make(map[string]providerRuntimeState)
-
-	r.mu.RLock()
-	providerTypes := make(map[core.Provider]string, len(r.providerTypes))
-	providerNames := make(map[core.Provider]string, len(r.providerNames))
-	maps.Copy(providerTypes, r.providerTypes)
-	maps.Copy(providerNames, r.providerNames)
-	r.mu.RUnlock()
-	configuredProviderModels, configuredProviderModelsMode := r.snapshotConfiguredProviderModels()
-
-	for _, provider := range providers {
-		providerName := providerNames[provider]
-		if providerName == "" {
-			providerName = providerTypes[provider]
-		}
-		if providerName == "" {
-			providerName = fmt.Sprintf("%p", provider)
-		}
-
-		configuredModels := configuredProviderModels[providerName]
-		resp, configuredReason, fetchAt, err := fetchProviderInventory(
-			ctx,
-			provider,
-			providerName,
-			providerTypes[provider],
-			configuredProviderModelsMode,
-			configuredModels,
-		)
-		var configuredUpstreamError string
-		if configuredReason != configuredProviderModelsNotApplied {
-			attrs := []any{
-				"provider", providerName,
-				"reason", string(configuredReason),
-				"configured_models", len(configuredModels),
-			}
-			if err != nil {
-				configuredUpstreamError = err.Error()
-				attrs = append(attrs, "error", err)
-				slog.Warn("upstream ListModels failed, using configured provider models", attrs...)
-			} else if configuredReason == configuredProviderModelsAllowlist {
-				slog.Debug("using configured provider models", attrs...)
-			} else {
-				slog.Warn("using configured provider models", attrs...)
-			}
-			err = nil
-		}
-		if err != nil {
-			slog.Warn("failed to fetch models from provider",
-				"provider", providerName,
-				"error", err,
-			)
-			failedProviders++
-			runtimeUpdates[providerName] = providerRuntimeState{
-				registered:          true,
-				lastModelFetchAt:    fetchAt,
-				lastModelFetchError: err.Error(),
-			}
-			continue
-		}
-
-		if resp == nil {
-			err := errors.New("provider returned nil model list")
-			slog.Warn("failed to fetch models from provider",
-				"provider", providerName,
-				"error", err,
-			)
-			failedProviders++
-			runtimeUpdates[providerName] = providerRuntimeState{
-				registered:          true,
-				lastModelFetchAt:    fetchAt,
-				lastModelFetchError: err.Error(),
-			}
-			continue
-		}
-
-		if len(resp.Data) == 0 {
-			err := errors.New("provider returned empty model list")
-			slog.Warn("provider returned empty model list",
-				"provider", providerName,
-			)
-			runtimeUpdates[providerName] = providerRuntimeState{
-				registered:          true,
-				lastModelFetchAt:    fetchAt,
-				lastModelFetchError: err.Error(),
-			}
-			if _, ok := newModelsByProvider[providerName]; !ok {
-				newModelsByProvider[providerName] = make(map[string]*ModelInfo)
-			}
-			continue
-		}
-
-		runtimeUpdate := providerRuntimeState{
-			registered:          true,
-			lastModelFetchAt:    fetchAt,
-			lastModelFetchError: configuredUpstreamError,
-		}
-		if configuredReason == configuredProviderModelsNotApplied {
-			runtimeUpdate.lastModelFetchSuccessAt = fetchAt
-		}
-		runtimeUpdates[providerName] = runtimeUpdate
-
-		if _, ok := newModelsByProvider[providerName]; !ok {
-			newModelsByProvider[providerName] = make(map[string]*ModelInfo, len(resp.Data))
-		}
-
-		for _, model := range resp.Data {
-			info := &ModelInfo{
-				Model:        model,
-				Provider:     provider,
-				ProviderName: providerName,
-				ProviderType: providerTypes[provider],
-			}
-			newModelsByProvider[providerName][model.ID] = info
-
-			if _, exists := newModels[model.ID]; exists {
-				// Model already registered by another provider, skip
-				// First provider wins for unqualified lookups.
-				slog.Debug("model already registered, skipping",
-					"model", model.ID,
-					"provider", providerName,
-					"owner", model.OwnedBy,
-				)
-				continue
-			}
-
-			newModels[model.ID] = info
-			totalModels++
-		}
-	}
-
-	if totalModels == 0 {
-		r.applyProviderRuntimeUpdates(runtimeUpdates)
-		if failedProviders == len(providers) {
-			return fmt.Errorf("failed to fetch models from any provider")
-		}
-		return fmt.Errorf("no models available: providers returned empty model lists")
-	}
-
-	// Enrich models with metadata from the model list (if loaded)
-	r.mu.RLock()
-	list := r.modelList
-	r.mu.RUnlock()
-	configOverrides := r.snapshotConfigOverrides()
-	metadataStats := metadataEnrichmentStats{}
-	if list != nil {
-		metadataStats = enrichProviderModelMaps(list, providerTypes, newModelsByProvider, nil)
-	}
-	metadataStats.Enriched += applyConfigMetadataOverrides(configOverrides, newModelsByProvider, nil)
-
-	// Atomically swap the models map and invalidate sorted caches
-	r.mu.Lock()
-	r.models = newModels
-	r.modelsByProvider = newModelsByProvider
-	r.applyProviderRuntimeUpdatesLocked(runtimeUpdates)
-	r.invalidateSortedCaches()
-	r.mu.Unlock()
-
-	// Mark as initialized
-	r.initMu.Lock()
-	r.initialized = true
-	r.initMu.Unlock()
-
-	attrs := []any{
-		"total_models", totalModels,
-		"providers", len(providers),
-		"failed_providers", failedProviders,
-	}
-	attrs = append(attrs, metadataStats.slogAttrs()...)
-	slog.Info("model registry initialized", attrs...)
-
-	return nil
-}
-
-func fetchProviderInventory(
-	ctx context.Context,
-	provider core.Provider,
-	providerName string,
-	providerType string,
-	mode config.ConfiguredProviderModelsMode,
-	configuredModels []string,
-) (*core.ModelsResponse, configuredProviderModelsApplyReason, time.Time, error) {
-	fetchAt := time.Now().UTC()
-	if mode == config.ConfiguredProviderModelsModeAllowlist && len(configuredModels) > 0 {
-		resp, reason := applyConfiguredProviderModels(
-			providerName,
-			providerType,
-			mode,
-			configuredModels,
-			nil,
-			nil,
-			fetchAt.Unix(),
-		)
-		return resp, reason, fetchAt, nil
-	}
-
-	resp, err := provider.ListModels(ctx)
-	fetchAt = time.Now().UTC()
-	resp, reason := applyConfiguredProviderModels(
-		providerName,
-		providerType,
-		mode,
-		configuredModels,
-		resp,
-		err,
-		fetchAt.Unix(),
-	)
-	return resp, reason, fetchAt, err
-}
-
-func (r *ModelRegistry) applyProviderRuntimeUpdates(updates map[string]providerRuntimeState) {
-	if len(updates) == 0 {
-		return
-	}
-
-	r.mu.Lock()
-	defer r.mu.Unlock()
-
-	r.applyProviderRuntimeUpdatesLocked(updates)
-}
-
-func (r *ModelRegistry) applyProviderRuntimeUpdatesLocked(updates map[string]providerRuntimeState) {
-	for providerName, update := range updates {
-		current := r.providerRuntime[providerName]
-		current.registered = update.registered || current.registered
-		if !update.lastModelFetchAt.IsZero() {
-			current.lastModelFetchAt = update.lastModelFetchAt
-		}
-		if !update.lastModelFetchSuccessAt.IsZero() {
-			current.lastModelFetchSuccessAt = update.lastModelFetchSuccessAt
-			if strings.TrimSpace(update.lastModelFetchError) == "" {
-				current.lastModelFetchError = ""
-			}
-		}
-		if strings.TrimSpace(update.lastModelFetchError) != "" {
-			current.lastModelFetchError = update.lastModelFetchError
-		}
-		r.providerRuntime[providerName] = current
-	}
-}
-
-// Refresh updates the model registry by fetching fresh model lists from providers.
-// This can be called periodically to keep the registry up to date.
-func (r *ModelRegistry) Refresh(ctx context.Context) error {
-	return r.Initialize(ctx)
-}
-
-func (r *ModelRegistry) acquireRefresh(ctx context.Context) (func(), error) {
-	if ctx == nil {
-		ctx = context.Background()
-	}
-	if err := ctx.Err(); err != nil {
-		return nil, registryRefreshAcquireError(err)
-	}
-	ch := r.refreshSemaphore()
-	select {
-	case ch <- struct{}{}:
-		return func() { <-ch }, nil
-	case <-ctx.Done():
-		return nil, registryRefreshAcquireError(ctx.Err())
-	}
-}
-
-func (r *ModelRegistry) refreshSemaphore() chan struct{} {
-	r.refreshOnce.Do(func() {
-		if r.refreshCh == nil {
-			r.refreshCh = make(chan struct{}, 1)
-		}
-	})
-	return r.refreshCh
-}
-
-func registryRefreshAcquireError(err error) *core.GatewayError {
-	if errors.Is(err, context.DeadlineExceeded) {
-		return core.NewProviderError("model_registry", http.StatusGatewayTimeout, "model registry refresh timed out before start", err)
-	}
-	return core.NewProviderError("model_registry", http.StatusRequestTimeout, "model registry refresh canceled before start", err)
-}
-
-// LoadFromCache loads the model list from the cache backend.
-// Returns the number of models loaded and any error encountered.
-func (r *ModelRegistry) LoadFromCache(ctx context.Context) (int, error) {
-	r.mu.RLock()
-	cacheBackend := r.cache
-	r.mu.RUnlock()
-
-	if cacheBackend == nil {
-		return 0, nil
-	}
-
-	modelCache, err := cacheBackend.Get(ctx)
-	if err != nil {
-		return 0, fmt.Errorf("failed to read cache: %w", err)
-	}
-
-	if modelCache == nil {
-		return 0, nil // No cache yet, not an error
-	}
-
-	// Build lookup maps from configured providers.
-	r.mu.RLock()
-	nameToProvider := make(map[string]core.Provider, len(r.providerNames))
-	nameToProviderType := make(map[string]string, len(r.providerNames))
-	providerOrderNames := make([]string, 0, len(r.providers))
-	for _, provider := range r.providers {
-		providerName := r.providerNames[provider]
-		if providerName == "" {
-			continue
-		}
-		providerOrderNames = append(providerOrderNames, providerName)
-	}
-	for provider, pName := range r.providerNames {
-		nameToProvider[pName] = provider
-		nameToProviderType[pName] = r.providerTypes[provider]
-	}
-	r.mu.RUnlock()
-
-	// Populate model maps from grouped cache structure. Unqualified lookups keep "first provider wins".
-	newModels := make(map[string]*ModelInfo)
-	newModelsByProvider := make(map[string]map[string]*ModelInfo)
-	cachedProviderTypes := make(map[string]string, len(modelCache.Providers))
-	for providerName, cachedProv := range modelCache.Providers {
-		provider, ok := nameToProvider[providerName]
-		if !ok {
-			// Provider not configured, skip all its models
-			continue
-		}
-		cachedProviderTypes[providerName] = strings.TrimSpace(cachedProv.ProviderType)
-		providerType := strings.TrimSpace(nameToProviderType[providerName])
-		if providerType == "" {
-			providerType = strings.TrimSpace(cachedProv.ProviderType)
-		}
-		providerModels := make(map[string]*ModelInfo, len(cachedProv.Models))
-		for _, cached := range cachedProv.Models {
-			info := &ModelInfo{
-				Model: core.Model{
-					ID:      cached.ID,
-					Object:  "model",
-					OwnedBy: cachedProv.OwnedBy,
-					Created: cached.Created,
-				},
-				Provider:     provider,
-				ProviderName: providerName,
-				ProviderType: providerType,
-			}
-			providerModels[cached.ID] = info
-			if _, exists := newModels[cached.ID]; !exists {
-				newModels[cached.ID] = info
-			}
-		}
-		newModelsByProvider[providerName] = providerModels
-	}
-
-	configuredProviderModels, configuredProviderModelsMode := r.snapshotConfiguredProviderModels()
-	if len(configuredProviderModels) > 0 {
-		for providerName, configuredModels := range configuredProviderModels {
-			provider, ok := nameToProvider[providerName]
-			if !ok {
-				continue
-			}
-			providerType := strings.TrimSpace(nameToProviderType[providerName])
-			if providerType == "" {
-				providerType = strings.TrimSpace(cachedProviderTypes[providerName])
-			}
-			providerModels := newModelsByProvider[providerName]
-			upstream := modelsResponseFromProviderMap(providerModels)
-			resp, reason := applyConfiguredProviderModels(providerName, providerType, configuredProviderModelsMode, configuredModels, upstream, nil, modelCache.UpdatedAt.Unix())
-			if reason == configuredProviderModelsNotApplied {
-				continue
-			}
-			newModelsByProvider[providerName] = modelInfoMapFromResponse(resp, provider, providerName, providerType)
-		}
-	}
-	newModels = rebuildGlobalModelMap(newModelsByProvider, providerOrderNames)
-
-	// Load model list data from cache if available
-	var list *modeldata.ModelList
-	if len(modelCache.ModelListData) > 0 {
-		parsed, parseErr := modeldata.Parse(modelCache.ModelListData)
-		if parseErr != nil {
-			slog.Warn("failed to parse cached model list data", "error", parseErr)
-		} else {
-			list = parsed
-		}
-	}
-
-	// Enrich cached models with model list metadata
-	metadataStats := metadataEnrichmentStats{}
-	if list != nil {
-		metadataStats = enrichProviderModelMaps(list, r.snapshotProviderTypes(), newModelsByProvider, nil)
-	}
-	configOverrides := r.snapshotConfigOverrides()
-	metadataStats.Enriched += applyConfigMetadataOverrides(configOverrides, newModelsByProvider, nil)
-
-	r.mu.Lock()
-	r.models = newModels
-	r.modelsByProvider = newModelsByProvider
-	r.invalidateSortedCaches()
-	if list != nil {
-		r.modelList = list
-		r.modelListRaw = modelCache.ModelListData
-	}
-	r.mu.Unlock()
-
-	attrs := []any{
-		"models", len(newModels),
-		"cache_updated_at", modelCache.UpdatedAt,
-	}
-	attrs = append(attrs, metadataStats.slogAttrs()...)
-	slog.Info("loaded models from cache", attrs...)
-
-	return len(newModels), nil
-}
-
-// SaveToCache saves the current model list to the cache backend.
-func (r *ModelRegistry) SaveToCache(ctx context.Context) error {
-	r.mu.RLock()
-	cacheBackend := r.cache
-	modelsByProvider := make(map[string]map[string]*ModelInfo, len(r.modelsByProvider))
-	for providerName, models := range r.modelsByProvider {
-		modelsByProvider[providerName] = make(map[string]*ModelInfo, len(models))
-		maps.Copy(modelsByProvider[providerName], models)
-	}
-	providerTypes := make(map[core.Provider]string, len(r.providerTypes))
-	maps.Copy(providerTypes, r.providerTypes)
-	modelListRaw := r.modelListRaw
-	r.mu.RUnlock()
-
-	if cacheBackend == nil {
-		return nil
-	}
-
-	mc := &modelcache.ModelCache{
-		UpdatedAt:     time.Now().UTC(),
-		Providers:     make(map[string]modelcache.CachedProvider, len(modelsByProvider)),
-		ModelListData: modelListRaw,
-	}
-
-	var totalModels int
-	for providerName, models := range modelsByProvider {
-		// Determine provider type and owned_by from any model in this provider group.
-		var pType, ownedBy string
-		for _, info := range models {
-			if ownedBy == "" {
-				ownedBy = info.Model.OwnedBy
-			}
-			if pType == "" {
-				pType = strings.TrimSpace(info.ProviderType)
-				if pType == "" {
-					pType = strings.TrimSpace(providerTypes[info.Provider])
-				}
-			}
-			if pType != "" && ownedBy != "" {
-				break
-			}
-		}
-		if pType == "" {
-			// No known provider type for this provider, skip entirely.
-			continue
-		}
-
-		modelIDs := make([]string, 0, len(models))
-		for modelID := range models {
-			modelIDs = append(modelIDs, modelID)
-		}
-		sort.Strings(modelIDs)
-
-		cachedModels := make([]modelcache.CachedModel, 0, len(modelIDs))
-		for _, modelID := range modelIDs {
-			info := models[modelID]
-			cachedModels = append(cachedModels, modelcache.CachedModel{
-				ID:      modelID,
-				Created: info.Model.Created,
-			})
-		}
-		mc.Providers[providerName] = modelcache.CachedProvider{
-			ProviderType: pType,
-			OwnedBy:      ownedBy,
-			Models:       cachedModels,
-		}
-		totalModels += len(cachedModels)
-	}
-
-	if err := cacheBackend.Set(ctx, mc); err != nil {
-		return fmt.Errorf("failed to save cache: %w", err)
-	}
-
-	slog.Debug("saved models to cache", "models", totalModels)
-	return nil
-}
-
-// InitializeAsync starts model fetching in a background goroutine.
-// It first loads any cached models for immediate availability, then refreshes from network.
-// Returns immediately after loading cache. The background goroutine will update models
-// and save to cache when network fetch completes.
-func (r *ModelRegistry) InitializeAsync(ctx context.Context) {
-	// First, try to load from cache for instant startup
-	cached, err := r.LoadFromCache(ctx)
-	if err != nil {
-		slog.Warn("failed to load models from cache", "error", err)
-	} else if cached > 0 {
-		slog.Info("serving traffic with cached models while refreshing", "cached_models", cached)
-	}
-
-	// Start background initialization
-	go func() {
-		initCtx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
-		defer cancel()
-
-		if err := r.Initialize(initCtx); err != nil {
-			slog.Warn("background model initialization failed", "error", err)
-			return
-		}
-
-		// Save to cache for next startup
-		if err := r.SaveToCache(initCtx); err != nil {
-			slog.Warn("failed to save models to cache", "error", err)
-		}
-	}()
-}
-
-// IsInitialized returns true if at least one successful network fetch has completed.
-// This can be used to check if the registry has fresh data or is only serving from cache.
-func (r *ModelRegistry) IsInitialized() bool {
-	r.initMu.Lock()
-	defer r.initMu.Unlock()
-	return r.initialized
-}
 
 // GetProvider returns the provider for the given model, or nil if not found
 func (r *ModelRegistry) GetProvider(model string) core.Provider {
@@ -1359,544 +803,4 @@ func (r *ModelRegistry) ProviderRuntimeSnapshots() []ProviderRuntimeSnapshot {
 	}
 
 	return result
-}
-
-// SetModelList stores the parsed model list and its raw bytes for cache persistence.
-func (r *ModelRegistry) SetModelList(list *modeldata.ModelList, raw json.RawMessage) {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	r.modelList = list
-	r.modelListRaw = raw
-}
-
-// EnrichModels re-applies model list metadata to all currently registered models.
-// Call this after SetModelList to update existing models with the new metadata.
-// Holds the write lock for the entire operation and replaces published ModelInfo
-// entries instead of mutating them in place so concurrent readers can safely keep
-// using older snapshots after unlocking.
-func (r *ModelRegistry) EnrichModels() {
-	_ = r.enrichModels()
-}
-
-func (r *ModelRegistry) enrichModels() metadataEnrichmentStats {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	return r.enrichModelsLocked()
-}
-
-func (r *ModelRegistry) enrichModelsLocked() metadataEnrichmentStats {
-	if len(r.models) == 0 {
-		return metadataEnrichmentStats{}
-	}
-	if r.modelList == nil && len(r.configMetadataOverrides) == 0 {
-		return metadataEnrichmentStats{}
-	}
-
-	providerTypes := make(map[core.Provider]string, len(r.providerTypes))
-	maps.Copy(providerTypes, r.providerTypes)
-
-	replacements := make(map[*ModelInfo]*ModelInfo, len(r.models))
-	stats := metadataEnrichmentStats{}
-	if r.modelList != nil {
-		stats = enrichProviderModelMaps(r.modelList, providerTypes, r.modelsByProvider, replacements)
-	}
-	stats.Enriched += applyConfigMetadataOverrides(r.configMetadataOverrides, r.modelsByProvider, replacements)
-	for modelID, info := range r.models {
-		if replacement, ok := replacements[info]; ok {
-			r.models[modelID] = replacement
-		}
-	}
-	r.invalidateSortedCaches()
-	return stats
-}
-
-func (r *ModelRegistry) setModelListAndEnrich(list *modeldata.ModelList, raw json.RawMessage) metadataEnrichmentStats {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	r.modelList = list
-	r.modelListRaw = raw
-	return r.enrichModelsLocked()
-}
-
-// ResolveMetadata resolves metadata for a model directly via the stored model list,
-// bypassing the registry key lookup. This handles cases where the usage DB stores
-// a response model ID (e.g., "gpt-4o-2024-08-06") that differs from the registry
-// key (e.g., "gpt-4o") by using the reverse index in the model list.
-func (r *ModelRegistry) ResolveMetadata(providerType, modelID string) *core.ModelMetadata {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	if r.modelList == nil {
-		return nil
-	}
-	return modeldata.Resolve(r.modelList, providerType, modelID)
-}
-
-// GetModelMetadata returns the metadata for a model, or nil if not found or not enriched.
-func (r *ModelRegistry) GetModelMetadata(modelID string) *core.ModelMetadata {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	if info, ok := r.models[modelID]; ok {
-		return info.Model.Metadata
-	}
-	return nil
-}
-
-// ResolvePricing returns the pricing metadata for a model, trying the registry first
-// and falling back to a reverse-index lookup via the model list.
-// Returns nil if no pricing is available.
-func (r *ModelRegistry) ResolvePricing(model, providerType string) *core.ModelPricing {
-	providerSelector := strings.TrimSpace(providerType)
-	if meta := r.getProviderModelMetadata(providerSelector, model); meta != nil && meta.Pricing != nil {
-		return meta.Pricing
-	}
-
-	meta := r.GetModelMetadata(model)
-	if meta != nil && meta.Pricing != nil {
-		return meta.Pricing
-	}
-	if providerSelector != "" {
-		meta = r.ResolveMetadata(r.metadataProviderType(providerSelector), r.metadataModelID(model))
-		if meta != nil && meta.Pricing != nil {
-			return meta.Pricing
-		}
-	}
-	return nil
-}
-
-func (r *ModelRegistry) getProviderModelMetadata(providerSelector, model string) *core.ModelMetadata {
-	providerSelector = strings.TrimSpace(providerSelector)
-	model = strings.TrimSpace(model)
-	if model == "" {
-		return nil
-	}
-
-	modelProviderName, modelID := splitModelSelector(model)
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-
-	if modelProviderName != "" {
-		if meta := metadataFromProviderModel(r.modelsByProvider[modelProviderName], modelID); meta != nil {
-			return meta
-		}
-		if r.hasConfiguredProviderNameLocked(modelProviderName) {
-			return nil
-		}
-	}
-
-	if providerSelector == "" {
-		return nil
-	}
-	if meta := metadataFromProviderModel(r.modelsByProvider[providerSelector], model); meta != nil {
-		return meta
-	}
-	return nil
-}
-
-func metadataFromProviderModel(providerModels map[string]*ModelInfo, model string) *core.ModelMetadata {
-	if len(providerModels) == 0 {
-		return nil
-	}
-	info := providerModels[strings.TrimSpace(model)]
-	if info == nil {
-		return nil
-	}
-	return info.Model.Metadata
-}
-
-func (r *ModelRegistry) metadataProviderType(providerSelector string) string {
-	providerSelector = strings.TrimSpace(providerSelector)
-	if providerSelector == "" {
-		return ""
-	}
-	if providerType := r.GetProviderTypeForName(providerSelector); providerType != "" {
-		return providerType
-	}
-	return providerSelector
-}
-
-func (r *ModelRegistry) metadataModelID(model string) string {
-	model = strings.TrimSpace(model)
-	providerName, modelID := splitModelSelector(model)
-	if providerName == "" {
-		return model
-	}
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	if r.hasConfiguredProviderNameLocked(providerName) {
-		return modelID
-	}
-	return model
-}
-
-// snapshotProviderTypes returns a copy of the providerTypes map for use outside the lock.
-func (r *ModelRegistry) snapshotProviderTypes() map[core.Provider]string {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	m := make(map[core.Provider]string, len(r.providerTypes))
-	maps.Copy(m, r.providerTypes)
-	return m
-}
-
-// snapshotConfigOverrides returns a copy of the configMetadataOverrides outer
-// and inner maps for use outside the lock. The inner *core.ModelMetadata
-// pointers are shared, which is safe because SetProviderMetadataOverrides
-// deep-clones on insertion and the registry never hands those values back out.
-func (r *ModelRegistry) snapshotConfigOverrides() map[string]map[string]*core.ModelMetadata {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	if len(r.configMetadataOverrides) == 0 {
-		return nil
-	}
-	out := make(map[string]map[string]*core.ModelMetadata, len(r.configMetadataOverrides))
-	for provider, inner := range r.configMetadataOverrides {
-		innerCopy := make(map[string]*core.ModelMetadata, len(inner))
-		for modelID, meta := range inner {
-			innerCopy[modelID] = meta
-		}
-		out[provider] = innerCopy
-	}
-	return out
-}
-
-func (r *ModelRegistry) snapshotConfiguredProviderModels() (map[string][]string, config.ConfiguredProviderModelsMode) {
-	r.mu.RLock()
-	defer r.mu.RUnlock()
-	mode := config.ResolveConfiguredProviderModelsMode(r.configuredProviderModelsMode)
-	if len(r.configuredProviderModels) == 0 {
-		return nil, mode
-	}
-	out := make(map[string][]string, len(r.configuredProviderModels))
-	for provider, models := range r.configuredProviderModels {
-		out[provider] = slices.Clone(models)
-	}
-	return out, mode
-}
-
-// collectionEmpty reports whether a reflect.Value representing a slice, array,
-// or map has no elements (covering both nil and non-nil-but-zero-length), and
-// falls back to reflect.Value.IsZero for other kinds. This lets override-
-// emptiness checks treat `modes: []` the same as an omitted field, which
-// IsZero alone would not.
-func collectionEmpty(v reflect.Value) bool {
-	switch v.Kind() {
-	case reflect.Slice, reflect.Map, reflect.Array:
-		return v.Len() == 0
-	}
-	return v.IsZero()
-}
-
-// structFieldsEmpty returns true if every field of the given struct value
-// passes collectionEmpty.
-func structFieldsEmpty(v reflect.Value) bool {
-	for i := 0; i < v.NumField(); i++ {
-		if !collectionEmpty(v.Field(i)) {
-			return false
-		}
-	}
-	return true
-}
-
-// metadataOverrideEmpty reports whether an override has no effective content.
-// An empty override (either nil or zero-valued on every field) would turn a
-// nil current metadata into a non-nil empty struct after MergeMetadata, so
-// callers should short-circuit on it. Uses reflect-based field inspection so
-// new fields on core.ModelMetadata are picked up automatically; Pricing is
-// handled separately so a non-nil pointer to an empty pricing block still
-// counts as empty.
-func metadataOverrideEmpty(m *core.ModelMetadata) bool {
-	if m == nil {
-		return true
-	}
-	if !pricingOverrideEmpty(m.Pricing) {
-		return false
-	}
-	tmp := *m
-	tmp.Pricing = nil
-	return structFieldsEmpty(reflect.ValueOf(tmp))
-}
-
-// pricingOverrideEmpty reports whether a pricing override has no effective
-// content — nil or every field at its zero value (with collections treated as
-// empty when length==0).
-func pricingOverrideEmpty(p *core.ModelPricing) bool {
-	if p == nil {
-		return true
-	}
-	return structFieldsEmpty(reflect.ValueOf(*p))
-}
-
-// applyConfigMetadataOverrides layers operator-declared metadata onto already-
-// enriched models. Call it after enrichProviderModelMaps with the same
-// replacements map (pass nil replacements for fresh, unpublished maps).
-// Returns the number of models whose metadata was updated.
-func applyConfigMetadataOverrides(
-	overrides map[string]map[string]*core.ModelMetadata,
-	modelsByProvider map[string]map[string]*ModelInfo,
-	replacements map[*ModelInfo]*ModelInfo,
-) int {
-	if len(overrides) == 0 {
-		return 0
-	}
-	// reverse lets us find the pre-enrichment pointer when an entry has
-	// already been replaced by enrichProviderModelMaps, so our replacement
-	// chain stays consistent from the caller's perspective. Always allocated
-	// when replacements is non-nil so the else-branch write below cannot hit
-	// a nil map when enrichment made no replacements.
-	var reverse map[*ModelInfo]*ModelInfo
-	if replacements != nil {
-		reverse = make(map[*ModelInfo]*ModelInfo, len(replacements))
-		for orig, repl := range replacements {
-			reverse[repl] = orig
-		}
-	}
-	applied := 0
-	for providerName, modelOverrides := range overrides {
-		providerModels, ok := modelsByProvider[providerName]
-		if !ok {
-			continue
-		}
-		for modelID, override := range modelOverrides {
-			if metadataOverrideEmpty(override) {
-				// A nil or effectively-empty override has nothing to
-				// contribute. Skipping avoids turning a nil current metadata
-				// into a non-nil empty struct, which the DeepEqual check
-				// below would not catch.
-				continue
-			}
-			current, ok := providerModels[modelID]
-			if !ok {
-				continue
-			}
-			merged := modeldata.MergeMetadata(current.Model.Metadata, override)
-			// Skip no-op merges so concurrent readers holding the current
-			// pointer keep a stable view when the override adds no new info.
-			if reflect.DeepEqual(current.Model.Metadata, merged) {
-				continue
-			}
-			if replacements == nil {
-				current.Model.Metadata = merged
-				applied++
-				continue
-			}
-			cloned := *current
-			cloned.Model.Metadata = merged
-			next := &cloned
-			providerModels[modelID] = next
-			if orig, hasOrig := reverse[current]; hasOrig {
-				replacements[orig] = next
-				reverse[next] = orig
-			} else {
-				replacements[current] = next
-				reverse[next] = current
-			}
-			applied++
-		}
-	}
-	return applied
-}
-
-func enrichProviderModelMaps(
-	list *modeldata.ModelList,
-	providerTypes map[core.Provider]string,
-	modelsByProvider map[string]map[string]*ModelInfo,
-	replacements map[*ModelInfo]*ModelInfo,
-) metadataEnrichmentStats {
-	if list == nil {
-		return metadataEnrichmentStats{}
-	}
-	stats := metadataEnrichmentStats{}
-	for _, providerModels := range modelsByProvider {
-		if len(providerModels) == 0 {
-			continue
-		}
-		stats.Providers++
-		accessor := &registryAccessor{
-			models:        providerModels,
-			providerTypes: providerTypes,
-			replacements:  replacements,
-		}
-		enrichStats := modeldata.Enrich(accessor, list)
-		stats.Enriched += enrichStats.Enriched
-		stats.Total += enrichStats.Total
-	}
-	return stats
-}
-
-// registryAccessor implements modeldata.ModelInfoAccessor.
-// The models map may be either an unpublished snapshot (Initialize, LoadFromCache)
-// or the live registry map (EnrichModels, which uses replacements to preserve
-// immutability of already-published ModelInfo values).
-type registryAccessor struct {
-	models        map[string]*ModelInfo
-	providerTypes map[core.Provider]string
-	replacements  map[*ModelInfo]*ModelInfo
-}
-
-func (a *registryAccessor) ModelIDs() []string {
-	ids := make([]string, 0, len(a.models))
-	for id := range a.models {
-		ids = append(ids, id)
-	}
-	return ids
-}
-
-func (a *registryAccessor) GetProviderType(modelID string) string {
-	info, ok := a.models[modelID]
-	if !ok {
-		return ""
-	}
-	if providerType := strings.TrimSpace(info.ProviderType); providerType != "" {
-		return providerType
-	}
-	return strings.TrimSpace(a.providerTypes[info.Provider])
-}
-
-func (a *registryAccessor) SetMetadata(modelID string, meta *core.ModelMetadata) {
-	if info, ok := a.models[modelID]; ok {
-		if a.replacements != nil {
-			cloned := *info
-			cloned.Model.Metadata = meta
-			replacement := &cloned
-			a.models[modelID] = replacement
-			a.replacements[info] = replacement
-			return
-		}
-		info.Model.Metadata = meta
-	}
-}
-
-// StartBackgroundRefresh starts a goroutine that periodically refreshes the model registry.
-// If modelListURL is non-empty, the model list is also re-fetched on each tick.
-// The returned stop function is blocking: it cancels the refresh loop and waits
-// for the goroutine to exit before returning, so callers should expect it to
-// block during shutdown until any in-flight refresh work unwinds.
-func (r *ModelRegistry) StartBackgroundRefresh(interval time.Duration, modelListURL string) func() {
-	ctx, cancel := context.WithCancel(context.Background())
-	done := make(chan struct{})
-	var stopOnce sync.Once
-
-	go func() {
-		defer close(done)
-		ticker := time.NewTicker(interval)
-		defer ticker.Stop()
-
-		for {
-			select {
-			case <-ctx.Done():
-				return
-			case <-ticker.C:
-				refreshCtx, refreshCancel := context.WithTimeout(ctx, 30*time.Second)
-				err := r.Initialize(refreshCtx)
-				refreshCancel()
-				if err != nil {
-					if !isBenignBackgroundRefreshError(ctx, err) {
-						slog.Warn("background model refresh failed", "error", err)
-					}
-				} else {
-					func() {
-						cacheCtx, cacheCancel := context.WithTimeout(ctx, 10*time.Second)
-						defer cacheCancel()
-						if err := r.SaveToCache(cacheCtx); err != nil {
-							if !isBenignBackgroundRefreshError(ctx, err) {
-								slog.Warn("failed to save models to cache after refresh", "error", err)
-							}
-						}
-					}()
-				}
-
-				// Also refresh model list if configured
-				if modelListURL != "" {
-					r.refreshModelList(ctx, modelListURL)
-				}
-			}
-		}
-	}()
-
-	return func() {
-		stopOnce.Do(func() {
-			cancel()
-			<-done
-		})
-	}
-}
-
-// RefreshModelList fetches the external model metadata list and re-enriches all
-// currently registered models. It does not persist the model cache; callers that
-// want durable startup data should call SaveToCache after this succeeds.
-func (r *ModelRegistry) RefreshModelList(ctx context.Context, url string) (int, error) {
-	if strings.TrimSpace(url) == "" {
-		return 0, nil
-	}
-
-	release, err := r.acquireRefresh(ctx)
-	if err != nil {
-		return 0, err
-	}
-	defer release()
-
-	models, _, err := r.refreshModelListLocked(ctx, url)
-	return models, err
-}
-
-func (r *ModelRegistry) refreshModelListLocked(ctx context.Context, url string) (int, metadataEnrichmentStats, error) {
-	list, raw, err := modeldata.Fetch(ctx, url)
-	if err != nil {
-		return 0, metadataEnrichmentStats{}, err
-	}
-	if list == nil {
-		return 0, metadataEnrichmentStats{}, nil
-	}
-
-	metadataStats := r.setModelListAndEnrich(list, raw)
-	return len(list.Models), metadataStats, nil
-}
-
-// refreshModelList fetches the model list and re-enriches all models.
-func (r *ModelRegistry) refreshModelList(ctx context.Context, url string) {
-	fetchCtx, cancel := context.WithTimeout(ctx, 45*time.Second)
-	defer cancel()
-
-	release, err := r.acquireRefresh(fetchCtx)
-	if err != nil {
-		if !isBenignBackgroundRefreshError(ctx, err) {
-			slog.Warn("failed to acquire model list refresh", "url", url, "error", err)
-		}
-		return
-	}
-	var (
-		models        int
-		metadataStats metadataEnrichmentStats
-	)
-	func() {
-		defer release()
-		models, metadataStats, err = r.refreshModelListLocked(fetchCtx, url)
-	}()
-	if err != nil {
-		if !isBenignBackgroundRefreshError(ctx, err) {
-			slog.Warn("failed to refresh model list", "url", url, "error", err)
-		}
-		return
-	}
-	if models == 0 {
-		return
-	}
-
-	if err := r.SaveToCache(fetchCtx); err != nil {
-		if !isBenignBackgroundRefreshError(ctx, err) {
-			slog.Warn("failed to save cache after model list refresh", "error", err)
-		}
-	}
-	attrs := []any{"models", models}
-	attrs = append(attrs, metadataStats.slogAttrs()...)
-	slog.Debug("model list refreshed", attrs...)
-}
-
-func isBenignBackgroundRefreshError(parent context.Context, err error) bool {
-	if err == nil {
-		return true
-	}
-	if parent == nil || parent.Err() == nil {
-		return false
-	}
-	return errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded)
 }
