+++ internal/providers/registry_init.go
@@ -1,0 +1,494 @@
+package providers
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"log/slog"
+	"maps"
+	"net/http"
+	"strings"
+	"sync"
+	"time"
+
+	"gomodel/config"
+	"gomodel/internal/core"
+	"gomodel/internal/modeldata"
+)
+
+// Initialize fetches models from all registered providers and populates the registry.
+// This should be called on application startup.
+func (r *ModelRegistry) Initialize(ctx context.Context) error {
+	release, err := r.acquireRefresh(ctx)
+	if err != nil {
+		return err
+	}
+	defer release()
+	return r.initialize(ctx)
+}
+
+func (r *ModelRegistry) initialize(ctx context.Context) error {
+	// Get a snapshot of providers with a read lock
+	r.mu.RLock()
+	providers := make([]core.Provider, len(r.providers))
+	copy(providers, r.providers)
+	r.mu.RUnlock()
+
+	// Build new model maps without holding the lock.
+	// This allows concurrent reads to continue using the existing map
+	// while we fetch models from providers (which may involve network calls).
+	newModels := make(map[string]*ModelInfo)
+	newModelsByProvider := make(map[string]map[string]*ModelInfo)
+	var totalModels int
+	var failedProviders int
+	runtimeUpdates := make(map[string]providerRuntimeState)
+
+	r.mu.RLock()
+	providerTypes := make(map[core.Provider]string, len(r.providerTypes))
+	providerNames := make(map[core.Provider]string, len(r.providerNames))
+	maps.Copy(providerTypes, r.providerTypes)
+	maps.Copy(providerNames, r.providerNames)
+	r.mu.RUnlock()
+	configuredProviderModels, configuredProviderModelsMode := r.snapshotConfiguredProviderModels()
+
+	for _, provider := range providers {
+		providerName := providerNames[provider]
+		if providerName == "" {
+			providerName = providerTypes[provider]
+		}
+		if providerName == "" {
+			providerName = fmt.Sprintf("%p", provider)
+		}
+
+		configuredModels := configuredProviderModels[providerName]
+		resp, configuredReason, fetchAt, err := fetchProviderInventory(
+			ctx,
+			provider,
+			providerName,
+			providerTypes[provider],
+			configuredProviderModelsMode,
+			configuredModels,
+		)
+		var configuredUpstreamError string
+		if configuredReason != configuredProviderModelsNotApplied {
+			attrs := []any{
+				"provider", providerName,
+				"reason", string(configuredReason),
+				"configured_models", len(configuredModels),
+			}
+			if err != nil {
+				configuredUpstreamError = err.Error()
+				attrs = append(attrs, "error", err)
+				slog.Warn("upstream ListModels failed, using configured provider models", attrs...)
+			} else if configuredReason == configuredProviderModelsAllowlist {
+				slog.Debug("using configured provider models", attrs...)
+			} else {
+				slog.Warn("using configured provider models", attrs...)
+			}
+			err = nil
+		}
+		if err != nil {
+			slog.Warn("failed to fetch models from provider",
+				"provider", providerName,
+				"error", err,
+			)
+			failedProviders++
+			runtimeUpdates[providerName] = providerRuntimeState{
+				registered:          true,
+				lastModelFetchAt:    fetchAt,
+				lastModelFetchError: err.Error(),
+			}
+			continue
+		}
+
+		if resp == nil {
+			err := errors.New("provider returned nil model list")
+			slog.Warn("failed to fetch models from provider",
+				"provider", providerName,
+				"error", err,
+			)
+			failedProviders++
+			runtimeUpdates[providerName] = providerRuntimeState{
+				registered:          true,
+				lastModelFetchAt:    fetchAt,
+				lastModelFetchError: err.Error(),
+			}
+			continue
+		}
+
+		if len(resp.Data) == 0 {
+			err := errors.New("provider returned empty model list")
+			slog.Warn("provider returned empty model list",
+				"provider", providerName,
+			)
+			runtimeUpdates[providerName] = providerRuntimeState{
+				registered:          true,
+				lastModelFetchAt:    fetchAt,
+				lastModelFetchError: err.Error(),
+			}
+			if _, ok := newModelsByProvider[providerName]; !ok {
+				newModelsByProvider[providerName] = make(map[string]*ModelInfo)
+			}
+			continue
+		}
+
+		runtimeUpdate := providerRuntimeState{
+			registered:          true,
+			lastModelFetchAt:    fetchAt,
+			lastModelFetchError: configuredUpstreamError,
+		}
+		if configuredReason == configuredProviderModelsNotApplied {
+			runtimeUpdate.lastModelFetchSuccessAt = fetchAt
+		}
+		runtimeUpdates[providerName] = runtimeUpdate
+
+		if _, ok := newModelsByProvider[providerName]; !ok {
+			newModelsByProvider[providerName] = make(map[string]*ModelInfo, len(resp.Data))
+		}
+
+		for _, model := range resp.Data {
+			info := &ModelInfo{
+				Model:        model,
+				Provider:     provider,
+				ProviderName: providerName,
+				ProviderType: providerTypes[provider],
+			}
+			newModelsByProvider[providerName][model.ID] = info
+
+			if _, exists := newModels[model.ID]; exists {
+				// Model already registered by another provider, skip
+				// First provider wins for unqualified lookups.
+				slog.Debug("model already registered, skipping",
+					"model", model.ID,
+					"provider", providerName,
+					"owner", model.OwnedBy,
+				)
+				continue
+			}
+
+			newModels[model.ID] = info
+			totalModels++
+		}
+	}
+
+	if totalModels == 0 {
+		r.applyProviderRuntimeUpdates(runtimeUpdates)
+		if failedProviders == len(providers) {
+			return fmt.Errorf("failed to fetch models from any provider")
+		}
+		return fmt.Errorf("no models available: providers returned empty model lists")
+	}
+
+	// Enrich models with metadata from the model list (if loaded)
+	r.mu.RLock()
+	list := r.modelList
+	r.mu.RUnlock()
+	configOverrides := r.snapshotConfigOverrides()
+	metadataStats := metadataEnrichmentStats{}
+	if list != nil {
+		metadataStats = enrichProviderModelMaps(list, providerTypes, newModelsByProvider, nil)
+	}
+	metadataStats.Enriched += applyConfigMetadataOverrides(configOverrides, newModelsByProvider, nil)
+
+	// Atomically swap the models map and invalidate sorted caches
+	r.mu.Lock()
+	r.models = newModels
+	r.modelsByProvider = newModelsByProvider
+	r.applyProviderRuntimeUpdatesLocked(runtimeUpdates)
+	r.invalidateSortedCaches()
+	r.mu.Unlock()
+
+	// Mark as initialized
+	r.initMu.Lock()
+	r.initialized = true
+	r.initMu.Unlock()
+
+	attrs := []any{
+		"total_models", totalModels,
+		"providers", len(providers),
+		"failed_providers", failedProviders,
+	}
+	attrs = append(attrs, metadataStats.slogAttrs()...)
+	slog.Info("model registry initialized", attrs...)
+
+	return nil
+}
+
+func fetchProviderInventory(
+	ctx context.Context,
+	provider core.Provider,
+	providerName string,
+	providerType string,
+	mode config.ConfiguredProviderModelsMode,
+	configuredModels []string,
+) (*core.ModelsResponse, configuredProviderModelsApplyReason, time.Time, error) {
+	fetchAt := time.Now().UTC()
+	if mode == config.ConfiguredProviderModelsModeAllowlist && len(configuredModels) > 0 {
+		resp, reason := applyConfiguredProviderModels(
+			providerName,
+			providerType,
+			mode,
+			configuredModels,
+			nil,
+			nil,
+			fetchAt.Unix(),
+		)
+		return resp, reason, fetchAt, nil
+	}
+
+	resp, err := provider.ListModels(ctx)
+	fetchAt = time.Now().UTC()
+	resp, reason := applyConfiguredProviderModels(
+		providerName,
+		providerType,
+		mode,
+		configuredModels,
+		resp,
+		err,
+		fetchAt.Unix(),
+	)
+	return resp, reason, fetchAt, err
+}
+
+func (r *ModelRegistry) applyProviderRuntimeUpdates(updates map[string]providerRuntimeState) {
+	if len(updates) == 0 {
+		return
+	}
+
+	r.mu.Lock()
+	defer r.mu.Unlock()
+
+	r.applyProviderRuntimeUpdatesLocked(updates)
+}
+
+func (r *ModelRegistry) applyProviderRuntimeUpdatesLocked(updates map[string]providerRuntimeState) {
+	for providerName, update := range updates {
+		current := r.providerRuntime[providerName]
+		current.registered = update.registered || current.registered
+		if !update.lastModelFetchAt.IsZero() {
+			current.lastModelFetchAt = update.lastModelFetchAt
+		}
+		if !update.lastModelFetchSuccessAt.IsZero() {
+			current.lastModelFetchSuccessAt = update.lastModelFetchSuccessAt
+			if strings.TrimSpace(update.lastModelFetchError) == "" {
+				current.lastModelFetchError = ""
+			}
+		}
+		if strings.TrimSpace(update.lastModelFetchError) != "" {
+			current.lastModelFetchError = update.lastModelFetchError
+		}
+		r.providerRuntime[providerName] = current
+	}
+}
+
+// Refresh updates the model registry by fetching fresh model lists from providers.
+// This can be called periodically to keep the registry up to date.
+func (r *ModelRegistry) Refresh(ctx context.Context) error {
+	return r.Initialize(ctx)
+}
+
+func (r *ModelRegistry) acquireRefresh(ctx context.Context) (func(), error) {
+	if ctx == nil {
+		ctx = context.Background()
+	}
+	if err := ctx.Err(); err != nil {
+		return nil, registryRefreshAcquireError(err)
+	}
+	ch := r.refreshSemaphore()
+	select {
+	case ch <- struct{}{}:
+		return func() { <-ch }, nil
+	case <-ctx.Done():
+		return nil, registryRefreshAcquireError(ctx.Err())
+	}
+}
+
+func (r *ModelRegistry) refreshSemaphore() chan struct{} {
+	r.refreshOnce.Do(func() {
+		if r.refreshCh == nil {
+			r.refreshCh = make(chan struct{}, 1)
+		}
+	})
+	return r.refreshCh
+}
+
+func registryRefreshAcquireError(err error) *core.GatewayError {
+	if errors.Is(err, context.DeadlineExceeded) {
+		return core.NewProviderError("model_registry", http.StatusGatewayTimeout, "model registry refresh timed out before start", err)
+	}
+	return core.NewProviderError("model_registry", http.StatusRequestTimeout, "model registry refresh canceled before start", err)
+}
+
+// InitializeAsync starts model fetching in a background goroutine.
+// It first loads any cached models for immediate availability, then refreshes from network.
+// Returns immediately after loading cache. The background goroutine will update models
+// and save to cache when network fetch completes.
+func (r *ModelRegistry) InitializeAsync(ctx context.Context) {
+	// First, try to load from cache for instant startup
+	cached, err := r.LoadFromCache(ctx)
+	if err != nil {
+		slog.Warn("failed to load models from cache", "error", err)
+	} else if cached > 0 {
+		slog.Info("serving traffic with cached models while refreshing", "cached_models", cached)
+	}
+
+	// Start background initialization
+	go func() {
+		initCtx, cancel := context.WithTimeout(context.Background(), 60*time.Second)
+		defer cancel()
+
+		if err := r.Initialize(initCtx); err != nil {
+			slog.Warn("background model initialization failed", "error", err)
+			return
+		}
+
+		// Save to cache for next startup
+		if err := r.SaveToCache(initCtx); err != nil {
+			slog.Warn("failed to save models to cache", "error", err)
+		}
+	}()
+}
+
+// IsInitialized returns true if at least one successful network fetch has completed.
+// This can be used to check if the registry has fresh data or is only serving from cache.
+func (r *ModelRegistry) IsInitialized() bool {
+	r.initMu.Lock()
+	defer r.initMu.Unlock()
+	return r.initialized
+}
+
+// StartBackgroundRefresh starts a goroutine that periodically refreshes the model registry.
+// If modelListURL is non-empty, the model list is also re-fetched on each tick.
+// The returned stop function is blocking: it cancels the refresh loop and waits
+// for the goroutine to exit before returning, so callers should expect it to
+// block during shutdown until any in-flight refresh work unwinds.
+func (r *ModelRegistry) StartBackgroundRefresh(interval time.Duration, modelListURL string) func() {
+	ctx, cancel := context.WithCancel(context.Background())
+	done := make(chan struct{})
+	var stopOnce sync.Once
+
+	go func() {
+		defer close(done)
+		ticker := time.NewTicker(interval)
+		defer ticker.Stop()
+
+		for {
+			select {
+			case <-ctx.Done():
+				return
+			case <-ticker.C:
+				refreshCtx, refreshCancel := context.WithTimeout(ctx, 30*time.Second)
+				err := r.Initialize(refreshCtx)
+				refreshCancel()
+				if err != nil {
+					if !isBenignBackgroundRefreshError(ctx, err) {
+						slog.Warn("background model refresh failed", "error", err)
+					}
+				} else {
+					func() {
+						cacheCtx, cacheCancel := context.WithTimeout(ctx, 10*time.Second)
+						defer cacheCancel()
+						if err := r.SaveToCache(cacheCtx); err != nil {
+							if !isBenignBackgroundRefreshError(ctx, err) {
+								slog.Warn("failed to save models to cache after refresh", "error", err)
+							}
+						}
+					}()
+				}
+
+				// Also refresh model list if configured
+				if modelListURL != "" {
+					r.refreshModelList(ctx, modelListURL)
+				}
+			}
+		}
+	}()
+
+	return func() {
+		stopOnce.Do(func() {
+			cancel()
+			<-done
+		})
+	}
+}
+
+// RefreshModelList fetches the external model metadata list and re-enriches all
+// currently registered models. It does not persist the model cache; callers that
+// want durable startup data should call SaveToCache after this succeeds.
+func (r *ModelRegistry) RefreshModelList(ctx context.Context, url string) (int, error) {
+	if strings.TrimSpace(url) == "" {
+		return 0, nil
+	}
+
+	release, err := r.acquireRefresh(ctx)
+	if err != nil {
+		return 0, err
+	}
+	defer release()
+
+	models, _, err := r.refreshModelListLocked(ctx, url)
+	return models, err
+}
+
+func (r *ModelRegistry) refreshModelListLocked(ctx context.Context, url string) (int, metadataEnrichmentStats, error) {
+	list, raw, err := modeldata.Fetch(ctx, url)
+	if err != nil {
+		return 0, metadataEnrichmentStats{}, err
+	}
+	if list == nil {
+		return 0, metadataEnrichmentStats{}, nil
+	}
+
+	metadataStats := r.setModelListAndEnrich(list, raw)
+	return len(list.Models), metadataStats, nil
+}
+
+// refreshModelList fetches the model list and re-enriches all models.
+func (r *ModelRegistry) refreshModelList(ctx context.Context, url string) {
+	fetchCtx, cancel := context.WithTimeout(ctx, 45*time.Second)
+	defer cancel()
+
+	release, err := r.acquireRefresh(fetchCtx)
+	if err != nil {
+		if !isBenignBackgroundRefreshError(ctx, err) {
+			slog.Warn("failed to acquire model list refresh", "url", url, "error", err)
+		}
+		return
+	}
+	var (
+		models        int
+		metadataStats metadataEnrichmentStats
+	)
+	func() {
+		defer release()
+		models, metadataStats, err = r.refreshModelListLocked(fetchCtx, url)
+	}()
+	if err != nil {
+		if !isBenignBackgroundRefreshError(ctx, err) {
+			slog.Warn("failed to refresh model list", "url", url, "error", err)
+		}
+		return
+	}
+	if models == 0 {
+		return
+	}
+
+	if err := r.SaveToCache(fetchCtx); err != nil {
+		if !isBenignBackgroundRefreshError(ctx, err) {
+			slog.Warn("failed to save cache after model list refresh", "error", err)
+		}
+	}
+	attrs := []any{"models", models}
+	attrs = append(attrs, metadataStats.slogAttrs()...)
+	slog.Debug("model list refreshed", attrs...)
+}
+
+func isBenignBackgroundRefreshError(parent context.Context, err error) bool {
+	if err == nil {
+		return true
+	}
+	if parent == nil || parent.Err() == nil {
+		return false
+	}
+	return errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded)
+}
